requirements.txt
asgiref==3.4.1
boto3==1.20.29
botocore==1.23.29
dj-database-url==0.5.0
Django==4.0.1
django-crispy-forms==1.13.0
django-filter==21.1
django-heroku==0.3.1
django-storages==1.12.3
gunicorn==20.1.0
jmespath==0.10.0
Pillow==9.0.0
psycopg2==2.9.3
python-dateutil==2.8.2
s3transfer==0.5.0
six==1.16.0
sqlparse==0.4.2
tzdata==2021.5
urllib3==1.26.7
whitenoise==5.3.0